refactor: ♻️ extract sidebar nav items into a composable
Moves the large `items` array out of SidebarLayout into useSidebarNavItems
so the component stays under the max-lines limit, and adds the Community
Credit entry. Behaviour unchanged; the spec router gains the new route.
