+++ app/src/components/ui/__tests__/SidebarLayout.spec.ts
@@ -119,6 +119,11 @@ describe('SidebarLayout.vue', () => {
           path: '/teams/:id/debt-financing/lender-marketplace',
           name: 'lender-marketplace',
           component: { template: '<div>Lender Marketplace</div>' }
+        },
+        {
+          path: '/teams/:id/community-credit',
+          name: 'community-credit',
+          component: { template: '<div>Community Credit</div>' }
         }
       ]
     })
